fix(opinion): show header on publicar page and serve fresh columnistas data
- Add layout.tsx to opinion/publicar so the Header renders for unauthenticated users who need the login button
- Replace cached API route + server action with a direct client-side WPGraphQL call in useOpinionAuthors — eliminates the 1-hour stale-data window that caused the Columnistas widget to show the second-to-last article
