+++ src/app/opinion/publicar/layout.tsx
@@ -1,0 +1,14 @@
+import { Header } from '@components/Header'
+
+export default function PublicarLayout({
+  children
+}: {
+  readonly children: React.ReactNode
+}) {
+  return (
+    <>
+      <Header />
+      {children}
+    </>
+  )
+}
